third_party/WebKit/LayoutTests/http/tests/devtools/sources/debugger-async/async-callstack-in-console.html

<html>
<head>
<script src="../../../inspector/inspector-test.js"></script>
<script src="../../../inspector/debugger-test.js"></script>
<script src="../../../inspector/console-test.js"></script>
<script>

function testFunction()
{
    debugger;
    console.clear();
    setTimeout(timeout1, 0);
}

function timeout1()
{
    console.trace();
    setTimeout(timeout2, 0);
}

function timeout2()
{
    setTimeout(timeout3, 0);
    throw Error("foo");
}

function timeout3()
{
    console.trace();
    var iframeWidnow = document.getElementById("iframe").contentWindow;
    tryPostMessage(iframeWidnow, "http://www.example.com");
}

function tryPostMessage(win, origin)
{
    try {
        win.postMessage("Trying origin=" + origin, origin);
    } catch(ex) {
        console.error("FAIL: Error sending message to " + origin + ". " + ex);
    }
}

var test = function() {
  var maxAsyncCallStackDepth = 8;
  var numberOfConsoleMessages = 5;

  SourcesTestRunner.setQuiet(true);
  SourcesTestRunner.startDebuggerTest(step1);

  function step1() {
    TestRunner.DebuggerAgent.setPauseOnExceptions(SDK.DebuggerModel.PauseOnExceptionsState.DontPauseOnExceptions);
    TestRunner.DebuggerAgent.setAsyncCallStackDepth(0);
    SourcesTestRunner.runTestFunctionAndWaitUntilPaused(step2);
  }

  function step2() {
    TestRunner.DebuggerAgent.setAsyncCallStackDepth(maxAsyncCallStackDepth).then(didPause);
  }

  function didPause() {
    ConsoleTestRunner.waitUntilNthMessageReceived(numberOfConsoleMessages, expandAndDumpConsoleMessages);
    SourcesTestRunner.resumeExecution();
  }

  function expandAndDumpConsoleMessages() {
    ConsoleTestRunner.expandConsoleMessages(dumpConsoleMessages);
  }

  function dumpConsoleMessages() {
    ConsoleTestRunner.dumpConsoleMessages(false, false, TestRunner.textContentWithLineBreaks);
    SourcesTestRunner.completeDebuggerTest();
  }
};

</script>
</head>

<body onload="runTest()">
<p>
Tests asynchronous call stacks printed in console.
</p>
<div><iframe src="../debugger/resources/post-message-listener.html"
    id="iframe" width="800" height="100" style="border: 1px solid black;">
</iframe></div>
</body>
</html>
