third_party/WebKit/LayoutTests/http/tests/devtools/sources/debugger-async/async-callstack-in-console-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests asynchronous call stacks printed in console.

Set timer for test function.
async-callstack-in-console.js:18 Console was cleared
async-callstack-in-console.js:24 console.trace

timeout1 @ async-callstack-in-console.js:24

setTimeout (async)

testFunction @ async-callstack-in-console.js:19
async-callstack-in-console.js:31 Uncaught Error: foo
    at timeout2 (async-callstack-in-console.js:31)

timeout2 @ async-callstack-in-console.js:31

setTimeout (async)

timeout1 @ async-callstack-in-console.js:25

setTimeout (async)

testFunction @ async-callstack-in-console.js:19
async-callstack-in-console.js:36 console.trace

timeout3 @ async-callstack-in-console.js:36

setTimeout (async)

timeout2 @ async-callstack-in-console.js:30

setTimeout (async)

timeout1 @ async-callstack-in-console.js:25

setTimeout (async)

testFunction @ async-callstack-in-console.js:19
async-callstack-in-console.js:44 Failed to execute 'postMessage' on 'DOMWindow': The target origin provided ('http://www.example.com') does not match the recipient window's origin ('http://127.0.0.1:8000').

tryPostMessage @ async-callstack-in-console.js:44

timeout3 @ async-callstack-in-console.js:38

setTimeout (async)

timeout2 @ async-callstack-in-console.js:30

setTimeout (async)

timeout1 @ async-callstack-in-console.js:25

setTimeout (async)

testFunction @ async-callstack-in-console.js:19
